+++ src/main/resources/templates/book/bookDetail.html
@@ -281,7 +281,7 @@
               <h2 class="book-section-title">리뷰</h2>
                 <div th:replace="~{book/review/register-block::review}"></div>
                 <div th:replace="~{book/review/list::reviews}"></div>
-                <div th:replace="~{fragments/page::page(url=${'/view/book/details/'+ book.id})}"></div>
+                <div th:replace="~{fragments/page::page}"></div>
             </div>
         </div>
     </div>
